👷 build("API: Environment Configuration): Patch setup and requirements with Pandas 2.0 dependencies
simplejson and delorean required to run pandas 2.0
Delorean is a library that provides easy and convenient DateTime conversions in Python. Dealing with generating lists of date-times, timezone shifting, and much more.
simplejson is a simple, fast, complete, correct and extensible JSON encoder and decoder for Python. It is pure Python code with no dependencies, but includes an optional C extension for a serious speed boost. simplejson exposes an API familiar to users of the standard library marshal and pickle module